Organise into Packages
Divide classes into packages according to their functions for better organisation
