+++ data/tasks.txt
@@ -1,3 +1,4 @@
 1 | hihihihihih
 0 | bathe | 2002-04-03
 0 | grow | 2004-03-05 | 2024-03-05
+0 | yes
