--- src/main/java/InvalidEditException.java
@@ -1,6 +1,0 @@
-public class InvalidEditException extends TalkaBotException {
-
-    public InvalidEditException(String edit) {
-        super("Sorry, what were you trying to " + edit + " again?");
-    }
-}
