--- src/main/java/InvalidScheduleException.java
@@ -1,6 +1,0 @@
-public class InvalidScheduleException extends TalkaBotException {
-
-    public InvalidScheduleException() {
-        super("Sorry, what task were you trying to schedule again?");
-    }
-}
