--- src/main/java/NoInputException.java
@@ -1,7 +1,0 @@
-public class NoInputException extends TalkaBotException {
-
-    public NoInputException() {
-        super("Oh no! Did you forget what you were going to say?");
-    }
-
-}
