--- src/main/java/TalkaBotException.java
@@ -1,6 +1,0 @@
-public class TalkaBotException extends RuntimeException {
-
-    public TalkaBotException(String message) {
-        super(message);
-    }
-}
