--- src/main/java/UnknownInputException.java
@@ -1,6 +1,0 @@
-public class UnknownInputException extends TalkaBotException {
-
-    public UnknownInputException(String str) {
-        super("\"" + str + "\"...? What does that mean?");
-    }
-}
