--- src/main/java/UnknownTimeException.java
@@ -1,6 +1,0 @@
-public class UnknownTimeException extends TalkaBotException {
-
-    public UnknownTimeException(String action) {
-        super("Please let me know when it " + action + " so I can record it for you!");
-    }
-}
