--- src/main/java/WrongTaskTypeException.java
@@ -1,6 +1,0 @@
-public class WrongTaskTypeException extends TalkaBotException {
-
-    public WrongTaskTypeException(String message) {
-        super("Sorry, this task " + message);
-    }
-}
