+++ src/main/java/talkabot/Parser.java
@@ -1,3 +1,7 @@
+package talkabot;
+
+import talkabot.exceptions.UnknownTimeException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
