+++ src/main/java/talkabot/exceptions/InvalidEditException.java
@@ -1,0 +1,8 @@
+package talkabot.exceptions;
+
+public class InvalidEditException extends TalkaBotException {
+
+    public InvalidEditException(String edit) {
+        super("Sorry, what were you trying to " + edit + " again?");
+    }
+}
