+++ src/main/java/talkabot/exceptions/NoInputException.java
@@ -1,0 +1,9 @@
+package talkabot.exceptions;
+
+public class NoInputException extends TalkaBotException {
+
+    public NoInputException() {
+        super("Oh no! Did you forget what you were going to say?");
+    }
+
+}
