+++ src/main/java/talkabot/exceptions/TalkaBotException.java
@@ -1,0 +1,8 @@
+package talkabot.exceptions;
+
+public class TalkaBotException extends RuntimeException {
+
+    public TalkaBotException(String message) {
+        super(message);
+    }
+}
