+++ src/main/java/talkabot/exceptions/UnknownInputException.java
@@ -1,0 +1,8 @@
+package talkabot.exceptions;
+
+public class UnknownInputException extends TalkaBotException {
+
+    public UnknownInputException(String str) {
+        super("\"" + str + "\"...? What does that mean?");
+    }
+}
