+++ src/main/java/talkabot/exceptions/WrongTaskTypeException.java
@@ -1,0 +1,8 @@
+package talkabot.exceptions;
+
+public class WrongTaskTypeException extends TalkaBotException {
+
+    public WrongTaskTypeException(String message) {
+        super("Sorry, this talkabot.task " + message);
+    }
+}
