+++ src/main/java/talkabot/task/Task.java
@@ -1,0 +1,35 @@
+package talkabot.task;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done talkabot.task with X
+    }
+
+    @Override
+    public String toString() {
+        return "[" + this.getStatusIcon() + "] " + this.description;
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsUndone() {
+        this.isDone = false;
+    }
+
+    public String fileString() {
+        return (this.isDone
+                ? "1 | "
+                : "0 | ")
+                + this.description;
+    }
+}
