+++ src/main/java/talkabot/task/ToDo.java
@@ -1,0 +1,13 @@
+package talkabot.task;
+
+public class ToDo extends Task {
+
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
